some shifting around + norm
