+++ src/weapon/weapon_draw.c
@@ -34,21 +34,19 @@ void	draw_weapon_loop_hook(void *arg)
 
 void	draw_muzzle_flash(t_window *window, mlx_texture_t *tex, bool draw)
 {
-	int				x_iter;
-	int				y_iter;
-	int				color;
+	int					x_iter;
+	int					y_iter;
+	int					color;
 	const t_vector_int	size = (t_vector_int){60, 40};
 	const t_vector_int	offset = (t_vector_int){510, 725};
 
 	y_iter = -size.y;
 	while (++y_iter < size.y)
 	{
-		if (y_iter + offset.y >= HEIGHT)
-			continue ;
 		x_iter = -size.x;
 		while (++x_iter < size.x)
 		{
-			if (x_iter + offset.x >= WIDTH)
+			if (x_iter + offset.x >= WIDTH || y_iter + offset.y >= HEIGHT)
 				continue ;
 			if (draw == true)
 				color = weapon_get_color(
@@ -64,9 +62,9 @@ void	draw_muzzle_flash(t_window *window, mlx_texture_t *tex, bool draw)
 
 void	draw_weapon(t_window *window, mlx_texture_t *tex)
 {
-	int				x_iter;
-	int				y_iter;
-	int				color;
+	int					x_iter;
+	int					y_iter;
+	int					color;
 	const t_vector_int	offset = (t_vector_int){WEAPON_OFFSET_X,
 		WEAPON_OFFSET_Y - window->player->weapon->cooldown};
 
@@ -127,51 +125,5 @@ void	draw_cross_hair(t_window *window)
 			if (abs(x_iter) > free_space)
 				mlx_put_pixel(window->hud->hud_img,
 					x_iter + WIDTH / 2, y_iter + HEIGHT / 2, CROSSHAIR_COLOR);
-	}
-}
-
-void	clean_weapon(t_window *window)
-{
-	int				x_iter;
-	int				y_iter;
-
-	y_iter = -WEAPON_SIZE_Y;
-	while (++y_iter < WEAPON_SIZE_Y)
-	{
-		if (y_iter + WEAPON_OFFSET_Y >= HEIGHT)
-			continue ;
-		x_iter = -WEAPON_SIZE_X;
-		while (++x_iter < WEAPON_SIZE_X)
-		{
-			if (x_iter + WEAPON_OFFSET_X >= WIDTH)
-				continue ;
-			mlx_put_pixel(window->hud->hud_img,
-				x_iter + WEAPON_OFFSET_X, y_iter + WEAPON_OFFSET_Y, 0x00000000);
-		}
-	}
-}
-
-void	clean_cross_hair(t_window *window)
-{
-	int			x_iter;
-	int			y_iter;
-	const int	size = 35 + WEAPON_COOLDOWN;
-	const int	width = 0;
-
-	y_iter = -size - 1;
-	while (y_iter++ <= size)
-	{
-		x_iter = -width - 1;
-		while (x_iter++ <= width)
-			mlx_put_pixel(window->hud->hud_img,
-				x_iter + WIDTH / 2, y_iter + HEIGHT / 2, 0x00000000);
-	}
-	y_iter = -width - 1;
-	while (y_iter++ <= width)
-	{
-		x_iter = -size - 1;
-		while (x_iter++ <= size)
-			mlx_put_pixel(window->hud->hud_img,
-				x_iter + WIDTH / 2, y_iter + HEIGHT / 2, 0x00000000);
 	}
 }
